Don't throw in TestApplicationComponentManager#clearState()
See https://github.com/google/dagger/issues/1988.

The issue is that in the try-catch block in MarkThatRulesRanRule

If anything in the "try" block throws an exception, then the call to clearState() in the "finally" block is guaranteed to also throw an exception, which will lose the original exception and makes debugging difficult.

Another approach here would be to store the original exception and throw it instead within the finally block.

This CL removes the Precondition check from clearState() which checked that the hiltTestRule had been set before clearing the state. This check was mostly just to prevent illegal usages where we tried to clear state before we set the state. However, removing it should be fairly safe because the method is pkg-private so we only use it ourselves.

RELNOTES=N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=323440332