Support BroadcastReceiver with @AndroidEntryPoint transform.
BroadcastReceiver#onReceive is abstract and the user cannot
write an implementation that calls super.onReceive since the
superclass is only updated after compilation. To workaround
this the AndroidEntryPoint transform will now check if the
@AndroidEntryPoint-annotated class has a marker field that
identifies it for the bytecode injection of a
super.onReceive call at the start of the code attribute
so that at runtime, injection is performed in the Hilt
generated class.

Fixes: https://github.com/google/dagger/issues/1918

RELNOTES=Support BroadcastReceiver with @AndroidEntryPoint transform.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=325484901