Avoid using ComponentDescriptors in AggregatedDepsProcessor.
This CL fixes a bug where an incorrect type given to @InstallIn failed in AggregatedDepsProcessor but pointed to the type rather than the element annotated with @InstallIn. For example, @InstallIn(ApplicationContext.class) pointed to ApplicationContext rather than the element annotated with the @InstallIn, which wasn't super helpful in finding the mistake.

The main issue was that the error message was coming from deep within the AggregatedDepsProcessor when we create ComponentDescriptors for each of the types. To create the component descriptor, we end up validating the @DefineComponent type but at that point we don't have context on where it was being used.

This CL moves the check further up in AggregatedDepsProcessor where we have more context, and fails if one of the classes passed to @InstallIn is not annotated with @DefineComponent.

In addition, AggregatedDepsProcessor shouldn't even be trying to create ComponentDescriptors, since it really only needs the TypeElement to write its metadata. So this CL also removes the reliance of AggregatedDepsProcessor on ComponentDescriptors.

RELNOTES=N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=325531579