Support an onCleared callback in the ActivityRetainedComponent.
This adds a default binding to the ActivityRetainedComponent
allowing users to register onCleared listeners that will be
invoked when the activity retained component and its instances
will be destroyed.

RELNOTES=Add support for receiving a callback for when the activity retained component will no longer be used and destroyed.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=326057949