Throw when a race is detected adding a OnClearedListener.
Its not supported to initialize in a background thread and schedule a a task/runnable to add a new OnClearedListener in the main thread. A possible race condition can happen when adding the new listeners after onCleared has been dispatched and the component is no longer needed. To prevent the possibility of weird bugs this might cause, this CL makes it so that addOnClearedListener throws an exception if its too late to add new listeners.

RELNOTES=Throw when a race is detected adding a OnClearedListener.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=326700301