[Hilt]: Add 'hilt-compiler' artifact to replace 'hilt-android-compiler'.
This artifact is identical to `hilt-android-compiler` except it has a different maven coordinate. We will be deprecating `hilt-android-compiler` in a future release.

The main purpose of this artifact name change is that it will be used for both android and non-android processing in the future.

RELNOTES=Added 'hilt-compiler' artifact to replace 'hilt-android-compiler'
PiperOrigin-RevId: 329928943