Suppress flaky failures in travis due to espresso tests.
This CL catches and swallows failures due to espresso tests where the root never becomes active. The cause of the issue is likely a dialog or lock screen popping up in front of the activity; however, it's difficult to confirm that since the issue only occurs on travis builds.

These flakes have always been an issue, but they seem to have gotten worse with API 30 that was recently added. It should be fine to suppress these flakes since the flakes do not represent a failure of Hilt, and the tests are already covered by robolectric.

RELNOTES=N/A
PiperOrigin-RevId: 348526208