Migrate HiltGradlePlugin to use AGP's new ASM pipeline for bytecode transformation.
Updates the HiltGradlePlugin to use AGP's new ASM pipeline if the developer is on AGP 4.2.0+, otherwise the plugin will use the older transform APIs. Using the new APIs allows for the plugin to not need a custom task to transform local tests as this is now supported with the new pipeline. If the developer is using AGP 4.2.0 then having 'enableTransformForLocalTests' is no longer needed and a warning will be shown with the hopes of migrating user away and once 4.2 is stable completely removing and deprecating the option. AGP's new transform pipeline is available in AGP 4.2.0 which is currently in beta.

Fixes: https://github.com/google/dagger/issues/2118
RELNOTES=Use AGP's new transform pipeline when developer is on AGP 4.2.0+.
PiperOrigin-RevId: 348818076