Fix for InstallException on continuous integration runs.
This CL attempts to fix one of our flaky continuous integration failures for instrumentation tests. The exception is com.android.ddmlib.InstallException, which is addressed in b/37018642 by extending the ADB_INSTALL_TIMEOUT. In this case, I've extended it from the default of 2 seconds up to 10 seconds for both kokoro and travis.

RELNOTES=N/A
PiperOrigin-RevId: 349335720