Prevent committing docs if there are no changes to avoid failing the travis build.
See https://travis-ci.org/github/google/dagger/builds/755270606

Before CL/352657712 util/generate-latest-docs.sh was allowed to fail silently because it was in the "after_success" section of  travis.yml. However, now it is in the "scripts" section which will fail the build. We could mark the job as "allow_failures" but it would be nice to catch other legit types of failures in this script.

RELNOTES=N/A
PiperOrigin-RevId: 352702083