Fix bug in generate-latest-docs.sh to prevent failure in travis.
This CL moves the "git diff-index --quiet HEAD --" call into the "if" statement. Previously, I was running the command first, and then trying to store it in a variable after, but due to "set -e" it was failing the script. By moving the call directly into the if statement we can avoid this.

RELNOTES=N/A
PiperOrigin-RevId: 353148115