Exclude Dagger local snapshots from Maven local repo cache.
This prevents the LOCAL-SNAPSHOT artifacts from being cached with the rest of the `.m2/repository` directory which in turn would cause other jobs to fail due to not using the local snapshots built for the workflow. Instead use the upload and download actions to share the artifacts between jobs.

RELNOTES=N/A
PiperOrigin-RevId: 354200073