Fix exit code check in generate-latest-docs.sh
I'm no Bash expert but it seems the combination of 'set -e' (exit script on nonzero exit command) does not work well with the vanilla command if check. So with this change I encapsulate the 'git diff-index' command in the compare command '[[ ]]' along with a concat of the exit code, which is then compared to '1' for 'there is changes in the index'.

RELNOTES=n/a
PiperOrigin-RevId: 354636089