Refactor SourceFileGenerator to allow for generating multiple types.
This could be helpful in some of our build performance experiments where we're generating multiple files in places where we used to be generating one.

I've also removed nameGeneratedType(), which was always just used to extract a package, which was slightly awkward. We never want a package that's different from the originating element, so I've used that exclusively.

RELNOTES=n/a
PiperOrigin-RevId: 360860252