Generate all test components in a common directory separate from the test class root if shareTestComponents flag is enabled. Before this change, the test component would only be generated in the shared directory if it was able to use the shared test components.
If only one root type exists in the compilation unit for a given simple name, that name is used to prefix the generated component name. If more than a one root shares a simple name (tests with the same name in separate packages), we have to disambiguate.

RELNOTES=N/A
PiperOrigin-RevId: 367252285