Perform Hilt aggregation in a Gradle task configured via the Hilt Gradle plugin.
This CL adds a new flag to the Hilt Plugin 'enableAggregatingTask' that will configure the project with a task that performs Hilt's aggregation mechanism (as opposed to being done in an aggregating annotation processor) and outputs Java sources with `@ComponentTreeDeps`. The generated sources are used as input of a Java compile task that will run Hilt's component generating processor and Dagger's ComponentProcessor, both isolating. The compiled classes are then added to the build pipeline via registerPostJavacGeneratedBytecode().

Classpath aggregation is configured by default in this mode but it is moved and only affects the aggregating and secondary compile tasks. This fixes issues with Lint and the 'androidDependencies' task. Furthermore aggregation is optimized via a transform that extracts only the aggregating packages out of library and dependency jars. The artifact transformation is cache and can be done in parallel and ultimately reduces the aggregated classpath size.

The logic for aggregating deps and generating `@ComponentTreeDeps` is temporary and does not match what the processor currently does, a follow-up change will be done to share the logic and avoid any discrepancy. Test component sharing is also turned ON by default in this mode. Furthermore, due to all processors being ISOLATING, this mode also helps with unit test build times (https://github.com/google/dagger/issues/2228) by avoiding generating unnecessary components.

Fixes: https://github.com/google/dagger/issues/2031
Fixes: https://github.com/google/dagger/issues/2288
Fixes: https://github.com/google/dagger/issues/2378
RELNOTES=Added a Hilt Gradle plugin flag 'enableAggregatingTask' that when enabled improves build performance by replacing Hilt's aggregating annotation processor by a Gradle task.
PiperOrigin-RevId: 373576325