Add the concept of ShardImplementation.
This CL is a no-op in terms of behavior and generated code. However, a follow-up CL will use ShardImplementations to extend sharding beyond just PrivateMethodBindingExpression. This will allow sharding in "default" mode, and will enable a fastInit optimization that should allow for ~20% smaller generated components and ~20% faster builds of the component.

Background:
Before this CL, a ComponentImplementation represented both the component implementation and the shard implementation. Calling `componentImplementation.getShard(key)` would return a possibly new instance of `ComponentImplementation` that represented a shard. This approach worked okay for sharding in fastInit because it was limited to only PrivateMethodBindingExpressions. However, when extending this to "default" mode the approach becomes error prone because it's too easy for a binding key to accidentally be located in multiple shards if we forget to call `componentImplementation.getShard(key)`.

This CL introduces ShardImplementation, which will replace most usages of ComponentImplemenation in our current code. Having a separate type for shards will force users to call `componentImplementation.getShard(key)` in order to pass the correct type.

RELNOTES=N/A
PiperOrigin-RevId: 374892777