Migrate ComponentProcessor's ProcessingStep to Step.
This is a follow-up to CL/375386035, which did the same things for dagger.android's processor.

Class usage was pretty pervasive in this processor, which is why this CL touches so many files. Where possible, I've tried to use ClassName in our own API, and only use String when interacting with the Step API.

RELNOTES=N/A
PiperOrigin-RevId: 376874775