Extend sharding to default mode.
This CL changes the BindingExpressions and CreationExpressions for a given binding to contribute to its corresponding shard. In addition, this CL removes the ability to contribute directly to a ComponentImplementation to force using a shard to contribute code.

The shard for a binding is calculated upfront using Tarjan's algorithm for calculating strongly connected components. This algorithm has two benefits 1) it groups the bindings in the graph into strongly connected components so that we can ensure that bindings participating in a cycle belong to the same shard and 2) it returns the groups in reverse topological order so that initializing Shard{i} before Shard{i+1} ensures all of Shard{i+1} dependencies have already been created.

RELNOTES=N/A
PiperOrigin-RevId: 377600402