Allow @AssistedFactory types to be injected as Provider<AssistedFactory>.
This is to allow situations where a JSR330-only library (like AutoFactory) may inject the factory as a Provider<> type. This is unnecessary since the factories are just shells around other providers, but if the library only codes to JSR330 then it may not be able to have knowledge of Dagger's AssistedFactory in order to not request a Provider of the factory. Other Dagger types like Lazy, Producer, Produced are still disallowed since they are Dagger types and so libraries using those types can reasonably add logic to request AssistedFactory types without the wrapper.

RELNOTES=n/a
PiperOrigin-RevId: 378450902