Do not transform invokespecial instructions produced for direct instantiations.
The instruction invokespecial is also used to invoke constructors and perform object instantiation. This CL fixes Hilt's transform to not change the owner of such instructions even if the owner is the older superclass since they are instantiating an object that happens to be the of the same type as the 'old' superclass.

Fixes https://github.com/google/dagger/issues/2662

RELNOTES=Fix an issue in Hilt's bytecode transform that would cause classes to fail validation if they contained an instantiation of an object whose type is the same as the superclass of the @AndroidEntryPoint annotated class.
PiperOrigin-RevId: 378500741