Make ComponentProcessor implement JavacBasicAnnotationProcessor.
This CL makes ComponentProcessor extend JavacBasicAnnotationProcessor, which can handle XProcessingStep and XProcessingEnv directly rather than having to convert it ourselves.

This CL also patches in the fixes to JavacBasicAnnotationProcessor, which is required for this change:
https://github.com/androidx/androidx/commit/8f65ad670a8c0650dca3567611ff8d8f13600117

At some point we will have a KspComponentProcessor that extends KspBasicAnnotationProcessor. At that point we'll likely want to refactor a lot of the code in ComponentProcessor into a separate class that can be shared by both KspComponentProcessor and ComponentProcessor, but for now I've kept it all in ComponentProcessor.

RELNOTES=N/A
PiperOrigin-RevId: 384769549