Fixes Dagger's shading to properly handle kotlin metadata
The issue was due to incorrect handling of shading in jarjar for kotlin metadata. In particular, jarjar does not replace UTF8 usages in the classfile, so the Strings in the kotlin metadata do not get replaced. This breaks when the class is compiled with kotlinc which uses the kotlin metadata during compilation.

To fix this, I've switched from jarjar to using Gradle's shadowJar to do the shading (https://imperceptiblethoughts.com/shadow/).

Also, I've added a kotlin regression test in gradle.

Closes #2765

RELNOTES=Fix #2765: Fixes Dagger's shading to properly handle kotlin metadata.
PiperOrigin-RevId: 387123329