Change build.gradle check to compare as Set.
Initially, I had assumed that the order of the returned sources was well defined (i.e. the order they appear in the source-tree) however, that doesn't seem to be the case because this check randomly started failing due to reverse order of the returned sources (https://github.com/google/dagger/runs/3622305048). Thus, this CL changes the check to compare the sources as Set rather than List.

RELNOTES=N/A
PiperOrigin-RevId: 397127271