Remove dead code in InjectionMethods.
This code has never been executed since its introduction in CL/164502837 (in 2017) because it is gated on an impossible conditional statement. We could correct this conditional to what it probably should have been, but this seems to just add unnecessary (and sometimes invalid) casts to the generated code, so I think it's best to just delete it.

RELNOTES=N/A
PiperOrigin-RevId: 398351534