Replace all unnecessary Class<?> references with ClassName in Dagger compiler code.
This CL replaces Class references with ClassName for all types that are only used to represent classes in generated code.

The issue with Class<?> references is that it can lead to unnecessary build dependencies in our compiler code (e.g. producers or android sdk) which can cause issues such as b/188559928.

RELNOTES=N/A
PiperOrigin-RevId: 401020296