Do not allow calls to earlySingletonComponent() if there are no @EarlyEntryPoints.
Before this change, this method would not throw if shareTestComponent was enabled, regardless of whether there were any @EarlyEntryPoints. After this change, EarlySingletonComponentCreatorImpl is no longer generated unless there are aggregated @EarlyEntryPoints. The ability to reflect on this class is what determines whether an exception is thrown.

RELNOTES=EarlyEntryPoints#get() will now throw if there are no @EarlyEntryPoint annotated interfaces.
PiperOrigin-RevId: 403191264