Update Gradle to 7.2 along with Kotlin to 1.5.31
Update usage of destinationDir and generatedSourcesDir to use the newer Property APIs as the old ones are deprecated. In Gradle 7.3 usage of the older APIs will cause a warning to be displayed to the user. To make sure we don't use Gradle deprecated APIs this CL also updates Gradle to latest stable and with -werror turned ON we can catch deprecated usages until newer Gradle versions are released.

Additionally to upgrade to Gradle 7.2 we also need to update Kotlin or otherwise there is a stdlib conflict due to Gradle bundling the Kotlin stdlib. The Kotlin update revealed deprecated API usages so those have to be cleaned up too, specifically String.capitalize from the stdlib text package, this is now replaced with a similar method in utils.

Finally, due to the Kotlin update a warning regarding mismatch JVM target version was revealed, this is fixed by declaring the Java source and target compatibility versions to match the Kotlin one.

Fixes https://github.com/google/dagger/issues/2967

RELNOTES=n/a
PiperOrigin-RevId: 406183362