Use reflection when getting the UnitTest component when AGP is 7.0+
This is needed because in 7.2 getUnitTest() ABI was broken as it returns a different type. By using reflection and upcasting to com.android.build.api.variant.Component we don't need to worry which UnitTest type is returned as we don't link to a specific method signature in the bytecode.

RELNOTES=Fix an AGP 7.2 incompatibility issue due to AGP breaking their API when cleaning up a deprecated type.
PiperOrigin-RevId: 407460171