Run Bazel tests in parallel with Gradle tests.
This CL splits Bazel build into a separate job from Bazel tests. This allows us to start Gradle tests after Bazel build rather than waiting on Bazel test. Bazel tests are now run in parallel to Gradle tests.

Note:I've used a separated cache for Bazel build and Bazel tests. This is because Bazel build's cache is much smaller (~200Mb compared to ~900Mb), which allows the loading/saving of the cache to be much faster. Although it takes up a bit more cache space to store these separately, the speed up is currently worth it given that Bazel build is on the critical path.

Sample Run:
https://github.com/google/dagger/actions/runs/1463107767

RELNOTES=N/A
PiperOrigin-RevId: 410045851