Move bazel-test in parallel with bazel-build.
This should drop our presubmit times from ~40min to ~20min
(presubmit run: https://github.com/google/dagger/actions/runs/1474082154)

Initially, I was hoping to share the building portion of bazel-build with bazel-test. However, the caching doesn't work. In particular, the cache will try to be restored from the current branch before it restores from the default branch. This means that currently, the bazel-test job never hits the bazel-test cache in the default branch because it hits the bazel-build cache in the current branch first. To fix this, I've moved these two jobs in parallel so that the bazel-test will get a chance to get a cache hit from the default branch's bazel-test cache, which not only will speed up the bazel-test job by a lot, but it also runs it 4-5min sooner than before.

RELNOTES=N/A
PiperOrigin-RevId: 410907070