Add better error message when SuperficialValidation throws an unexpected exception.
Note: The validateType(s) API now takes an extra String to pass in a description of what the type is (e.g. superclass, interface, etc.) since otherwise the error message would lose track of how the type is related to the element that is being validated. The validateElement(s) API does not take this parameter, because using its ElementKind should be enough context to piece together what actually failed.

An example of the new error message is shown below:

```
Validation trace:
    => SuperficialInjectValidator.validate: android.content.BroadcastReceiver
    => CLASS element: android.content.BroadcastReceiver
    => CLASS element: android.content.BroadcastReceiver.PendingResult
    => FIELD element: mType
    => annotation: @android.compat.annotation.UnsupportedAppUsage(maxTargetSdk=28, trackingBug=115609023L)
    => annotation value: maxTargetSdk
    => 'int' annotation value, 28, with expected type: <nulltype>
```

RELNOTES=Add better error message when SuperficialValidation throws an unexpected exception
PiperOrigin-RevId: 416097069