Replace SuperficialInjectValidator with inline validation.
Inlining the superficial validation should make it easier to maintain since we don't have to keep SuperficialInjectValidation in sync with InjectValidator. It also makes it easier to reason about why we are validating/not-validating certain types since the superficial validation is directly inlined to where we use it in the InjectValidator.

RELNOTES=N/A
PiperOrigin-RevId: 420064963