Add a repro test for issue #3163.
This CL adds a repro test for a provides method that references a transitive type with type parameters that is missing from the classpath.

Currently, the test just reproduces the issue described in #3163. A fix may take a bit of work, so that will be done in follow-up CLs.

RELNOTES=N/A
PiperOrigin-RevId: 421659953