Fix missing scopes and qualifiers on @Provides methods.
This CL fixes two separate issue of @Provides qualifiers. First, the build now fails when a qualifier type is not valid rather than just ignoring the qualifier. Second, it avoids validating non-qualifier types in upstream builds so that we avoid a regression (i.e. we only cause a failure when something is actually broken).

This CL currently only fixes @Provides. It does not address @Binds, @Produces, etc. Those will be addressed in a follow-up CL.

RELNOTES=Fixes missing scopes and qualifiers on @Provides methods.
PiperOrigin-RevId: 424441787