Update the transitive-annotation-app example to test the case of a valid annotation type with invalid annotation value.
This CL makes the following changes to the example app

```
// Old annotations
@MySimpleTransitiveAnnotation
@MyTransitiveAnnotation(VALUE)

// New annotations
@MyTransitiveAnnotation
@MyAnnotation(MyTransitiveType.VALUE)
```

Besides renaming some annotations, the important part of this change is that `MyAnnotation` is not a transitive type. This change is useful because it allows us to test the case where the annotation type itself is valid but the annotation value is not.

RELNOTES=N/A
PiperOrigin-RevId: 427584063