Fix bug in DependencyRequestValidator#missingQualifierMetadata
This CL fixes a bug introduced in CL/406404317. In particular, the if-conditional logic (`!isTypeElement(fieldElement.getEnclosingElement())`) makes it possible to accidentally get a class cast exception later (when calling `asTypeElement(fieldElement.getEnclosingElement())`).

AFAIK, this hasn't caused any actual issues with clients, likely because it would fail earlier with a different error message.

RELNOTES=N/A
PiperOrigin-RevId: 431055748