Allow Hilt view constructor to contain non-declared types.
This CL fixes an issue in ViewGenerator#isRestrictedApiConstructor() where we were assuming that the parameter type could be represented by a TypeElement. To fix this, we now first check that the parameter type is a TypeKind.DECLARED type before getting a TypeElement for it.

Fixes #3296

RELNOTES=Allow Hilt view constructor to contain non-declared types.
PiperOrigin-RevId: 434522808