Use a shard class for all bindings.
This CL, changes the generated code such that all implementation for binding creation is put into component shards rather than the component itself.

This change is motivated by build performance issues found in javac for complex user @Component-annotated interfaces. In some particularly bad cases, we've been able to reduce the build time of the component by up to 75% with this change.

RELNOTES=N/A
PiperOrigin-RevId: 440203238