Move Hilt's field/method descriptor logic into separate class.
This CL refactors Hilt's field/method descriptor logic out of KotlinMetadata and into a separate ElementDescriptors class. This allows BroadcastReceiverGenerator to share the logic and remove its dependency on DaggerElements, which will be removed during the XProcessing migration.

This CL also forks over Dagger's ElementDescriptorsTest. The Dagger version of this test will be removed once DaggerElements is removed.

RELNOTES=N/A
PiperOrigin-RevId: 444601570