Fix O(N^2) performance issue when determining declaration order.
Note: This CL fixes the issue for `InjectionSiteFactory`, but I ended up just inlining the last remaining (unfixed) usage into `DiagnosticMessageGenerator`.

Looks like there's a number of similar comparators within `DiagnosticMessageGenerator`, which we may want to eventually look into fixing them all at some point. On the other hand, the `DiagnosticMessageGenerator` only runs when there's errors so it's likely not as big of an issue.

RELNOTES=N/A
PiperOrigin-RevId: 444957049