Add Room's room-compiler-processing-testing jar into Dagger.
This jar is included to run Kotlin compilation tests using the Room testing infrastructure. This jar is included manually instead of depending on the artifact published in Maven to reduce iteration time as new changes are made to the compiler testing APIs. Once that has stabilized, this will be switched to depend on the official Maven artifact.

Also update the regular Room artifact to keep it at the same version.

RELNOTES=n/a
PiperOrigin-RevId: 448118701