Migrate `XAnnotations.toString()` and `MoreAnnotationMirrors.toStableString()` to XProcessing.
Note that `XAnnotations.toString()` previously used `AnnotationMirrors.toString()` under the hood, which happened to produced a stable string representation of the annotation. However, a stable string representation is not really what `AnnotationMirrors.toString()` was created for -- it was created to be used as valid java source code (see javadoc https://javadoc.io/static/com.google.auto/auto-common/1.2/com/google/auto/common/AnnotationMirrors.html#toString-javax.lang.model.element.AnnotationMirror-).

Thus, to be more deliberate/explicit about our String usages of annotations, which is mostly to report error messages, I've replaced all `XAnnotations.toString()` with `XAnnotations.toStableString()`. This should make it clearer that we're not using this String for use in source code and we only need the string to be stable.

RELNOTES=N/A
PiperOrigin-RevId: 465165919